Allow ORT to be disabled with NOORT flag
