+++ hugot.go
@@ -19,7 +19,7 @@ type Session struct {
 	environmentDestroy              func() error
 }
 
-func newSession(runtime string, opts ...options.WithOption) (*Session, error) {
+func newSession(runtime string, additionalSetup func(*Session) (*Session, error), opts ...options.WithOption) (*Session, error) {
 
 	parsedOptions := options.Defaults()
 	parsedOptions.Runtime = runtime
@@ -47,7 +47,7 @@ func newSession(runtime string, opts ...options.WithOption) (*Session, error) {
 		// No session specific initialisation for these currently
 		return session, nil
 	case "ORT":
-		return ortSession(session)
+		return additionalSetup(session)
 	default:
 		return nil, errors.New("unsupported runtime: " + parsedOptions.Runtime)
 	}
